requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=1.10.7
pytest>=7.3.1
black>=23.3.0
ruff>=0.0.267
mypy>=1.3.0
firecrawl-py
anthropic
mcp
litellm
